Read players' move. Put mark on the board. Validate players' input and handle errors.
Method #ask_for_move ask players to enter coordinates of position where
they want to put they sign and return them to the block transfered to
Game#move method to put mark on a board. When user entered invalid
coordinates or position is not available for move (occupied), it keeps asking
for coordinates till they are valid and position is free for move.

Method #validate_move_coordinate typecast given coordinate to Integer
and in case or errors (TypeError, ArgumentError) rethrow them as
ValidateError while Ruby cares about chaining original error to
rethrowed as a cause(available to retrieval through a call to #cause
method on ValidError exception).
When it ensured that valid Integer value was given it checks that
entered coordinate belongs to the board, otherwise it raises different
type of ValidateError exception.

As a design decision board is not transfered to #ask_for_move method,
but instead block with check for position availability is provided.
This is allow to keep knowledge about game entities limited Cli#call
method without spreading them around. This is should simplify future
refactorings.

Changes:
- Introduce and implement new methods:
  - ::TickTackToe::Cli#ask_for_move
  - ::TickTackToe::Cli#validate_move_coordinate